Move tests to new API
This includes moving the sample tests to Topiary cli (because topiary no longer knows about languages).

Additionally, bash was reverted back to an older version due to a regression that needs to be investigated.
