+++ topiary-cli/tests/samples/expected/bash.sh
