+++ topiary-cli/tests/samples/expected/tree-sitter-query.scm
