+++ topiary-cli/tests/samples/input/bash.sh
